+++ forest/__init__.py
@@ -6,6 +6,12 @@ forecasts alongside observations.
 
 .. automodule:: forest.redux
 
+.. automodule:: forest.rx
+
+.. automodule:: forest.observe
+
+.. automodule:: forest.series
+
 .. automodule:: forest.keys
 
 .. automodule:: forest.config